+++ backend/canisters/user/impl/src/model/direct_chat.rs
@@ -23,6 +23,7 @@ pub struct DirectChat {
 impl DirectChat {
     pub fn new(
         them: UserId,
+        my_user_id: UserId,
         user_type: UserType,
         events_ttl: Option<Milliseconds>,
         anonymized_chat_id: u128,
@@ -31,7 +32,7 @@ impl DirectChat {
         DirectChat {
             them,
             date_created: now,
-            events: ChatEvents::new_direct_chat(them, events_ttl, anonymized_chat_id, now),
+            events: ChatEvents::new_direct_chat(them, my_user_id, events_ttl, anonymized_chat_id, now),
             unread_message_index_map: UnreadMessageIndexMap::default(),
             read_by_me_up_to: Timestamped::new(None, now),
             read_by_them_up_to: Timestamped::new(None, now),