+++ backend/canisters/user/impl/src/updates/c2c_user_canister.rs
@@ -151,10 +151,13 @@ fn process_event(event: UserCanisterEvent, caller_user_id: UserId, state: &mut R
         }
         UserCanisterEvent::SetEventsTtl(args) => {
             let is_new_chat = !state.data.direct_chats.exists(&caller_user_id.into());
-            let chat = state
-                .data
-                .direct_chats
-                .get_or_create(caller_user_id, UserType::User, || state.env.rng().random(), now);
+            let chat = state.data.direct_chats.get_or_create(
+                caller_user_id,
+                UserType::User,
+                state.env.canister_id().into(),
+                || state.env.rng().random(),
+                now,
+            );
 
             let last_updated_timestamp = chat.events.get_events_time_to_live().timestamp;