+++ backend/canisters/user/impl/src/updates/start_video_call.rs
@@ -108,10 +108,13 @@ pub fn handle_start_video_call(
         sender_context: None,
     };
 
-    let chat = state
-        .data
-        .direct_chats
-        .get_or_create(other, UserType::User, || state.env.rng().random(), now);
+    let chat = state.data.direct_chats.get_or_create(
+        other,
+        UserType::User,
+        state.env.canister_id().into(),
+        || state.env.rng().random(),
+        now,
+    );
 
     let mute_notification = their_message_index.is_some() || chat.notifications_muted.value;