+++ backend/libraries/chat_events/src/chat_events_list.rs
@@ -1172,7 +1172,13 @@ mod tests {
         let memory = MemoryManager::init(DefaultMemoryImpl::default());
         stable_memory_map::init_with_small_entries_map(memory.get(MemoryId::new(1)), memory.get(MemoryId::new(2)));
 
-        let mut events = ChatEvents::new_direct_chat(Principal::from_slice(&[1]).into(), events_ttl, random(), 1);
+        let mut events = ChatEvents::new_direct_chat(
+            Principal::from_slice(&[1]).into(),
+            Principal::from_slice(&[2]).into(),
+            events_ttl,
+            random(),
+            1,
+        );
 
         push_events(&mut events, 2);